requirements-test.txt
# Test dependencies
-r requirements.txt

# PyTorch (CPU) wheels live on the official PyTorch index.
# This keeps PyPI as the primary index while allowing torch to resolve.
--extra-index-url https://download.pytorch.org/whl/cpu

torch>=2.1

# ONNX export + ONNXRuntime parity tests (CPU).
onnx>=1.14
onnxruntime>=1.17

# Optional (used by tools/rtdetr_pose train_minimal when --real-images enabled)
Pillow>=10.0

# Optional: lets you run `pytest` even though tests use `unittest`.
pytest>=7.0

# ONNXRuntime is used for CPU-friendly PyTorch ↔ ONNX parity tests.
onnxruntime>=1.16

# Required by PyTorch 2.10+ for ONNX export
onnxscript>=0.1
